Omit pulses that are not enabled
Unlike steps, the user had no direct access to the pulse object so we don't have to create them
if we know at creation time that they are statically disabled.

There is still a side-effect from adding the channel though, namely that the channel is enabled
for the sequence. The optimization can only be done after the channel has been added
to preserve this side-effect.